[ratzdab] improve conversion of rat mc to zdab
If tracking is available, we add the entire track structure to the ZDAB
record instead of just the primary MCParticles. In this case, we flatten
the RAT track structure and resolve the parent/child relationships using
a fuzzy position matching like that in the RAT TrackNav. In either case,
made the following improvements:

* Unit conversions for distances
* Fix representation of momenta (direction cosines)
* Convert PDG code to SNOMAN particle code
* Convert KE to total energy, via a new mass table

Still yet to do is the interaction type code, which will require
improved accounting in RAT to be useful.