Tidying things up. Added a README to keep track of versions. Having issues with the March2016 new PMTINFO - some channels have 0x13 type which is not recognised so output is 5 PMTs short. Have also had to modify readin method to allow universal newline mode as this new file seems to be in slightly different format (added U to readin method)
Use PMT_INFO_Dec2014.csv (renamed) as default input.
In both these files have hand editted pmt 18/15/0 to be type 0x00 (was 0x40 BUTT) see message from Noel:
---
This particular channel is somewhat special and there are other special channels. You should also be able to find a bunch of BUTTs at the center of the detector (in reality these BUTTs have long disappeared) -  I guess what I'm saying is that there is some cleaning up that needs to happen.
So, in the case of 18/15/0: this used to be a NECK PMT, so It looks like I reset that channel location to an OWL location after removing the neck tube years ago. So yes, the tube type should probably be set to 0x00.
---
Need to double check Neck positions but have implemented these.
Have commented out is_physical as this is redundant info, but new output has linked_lcn in it.